+++ CHANGELOG.md
@@ -5,8 +5,9 @@
 BREAKING CHANGES:
 
 - The deprecated `contact_id` field has been removed from the `dnsimple_lets_encrypt_certificate` resource. (dnsimple/terraform-provider-dnsimple#93)
-- The `id` and `domain_id` fields on the `dnsimple_lets_encrypt_certificate` are now of type `int64` instead of `string`. (dnsimple/terraform-provider-dnsimple#93)
+- The `id` field on the `dnsimple_lets_encrypt_certificate` is now of type `int64` instead of `string` to keep in line with the API. (dnsimple/terraform-provider-dnsimple#93)
 - The `domain_id` field on the `dnsimple_lets_encrypt_certificate` is now required. (dnsimple/terraform-provider-dnsimple#93)
+- The `expires_on` attribute on the `dnsimple_lets_encrypt_certificate` has been renamed to `expires_at` to keep in line with the API. (dnsimple/terraform-provider-dnsimple#93)
 
 ## 0.16.3